found another bug in client
Scot was accidentally taking the first part of a message instead of the
second part so an error was being thrown for valid cell name. I think
now that those two little bugs are fixed it works for the most part.
Note: there is a problem where you can crash the client by making the
window really small.