acutal problem
I thought about and realized you should not need to go backwards that
was not the bug. It had something to do with parsing and normalizing. We
were normalizing contents instead of parsed and parsed was always off.
So I changed it so parsed now starts with content added back to the
token parser. The normalization code may now be redundant in parts